relocate single-service enums into their service modules
Move ResetType → services/ecu_reset.rs, DiagnosticSessionType →
services/diagnostic_session_control.rs, SecurityAccessType →
services/security_access.rs, CommunicationControlType + CommunicationType →
services/communication_control.rs, DtcSettings → services/control_dtc_settings.rs,
RoutineControlSubFunction → services/routine_control.rs.

Remove the now-empty shared/ source files. Update shared/mod.rs to drop
the five moved modules. Wire all enums through services::mod.rs and re-export
them flat from the crate root via pub use services::{...}; public paths unchanged.
